requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
git+https://github.com/insom/flask-coin.git#egg=Flask-Coin
